feat: implement network runtime support and integrate net_rt into the native compiler

+++ src/runtime/net_rt.h
@@ -12,17 +12,17 @@ void vit_net_init(void);
 void vit_net_cleanup(void);
 
 // Socket Primitives
-int vit_net_socket_create(void);
-int vit_net_socket_bind(int fd, const char* host, int port);
-int vit_net_socket_listen(int fd, int backlog);
-int vit_net_socket_accept(int fd);
-int vit_net_socket_connect(int fd, const char* host, int port);
-int vit_net_socket_send(int fd, const char* data, int len);
-int vit_net_socket_recv(int fd, char* buf, int max_len);
-void vit_net_socket_close(int fd);
+double vit_net_socket_create(void);
+double vit_net_socket_bind(double fd, const char* host, double port);
+double vit_net_socket_listen(double fd, double backlog);
+double vit_net_socket_accept(double fd);
+double vit_net_socket_connect(double fd, const char* host, double port);
+double vit_net_socket_send(double fd, const char* data, double len);
+double vit_net_socket_recv(double fd, char* buf, double max_len);
+void vit_net_socket_close(double fd);
 
 // Helper Utilities
-char* vit_net_recv_string(int fd, int max_len);
+char* vit_net_recv_string(double fd, double max_len);
 
 #ifdef __cplusplus
 }
